Revert "Adding component for Video page"
