--- content/videos.json
@@ -1,7 +1,0 @@
-{
-  "videos": [
-    {
-      "videoId": "az-ILjdFCe8"
-    }
-  ]
-}
